Separate Jetty/Tomcat plugins
